+++ src/constants.ts
@@ -92,6 +92,7 @@ export const enum Commands {
     JiraLogin = 'atlascode.jira.login',
     JiraAPITokenLogin = 'atlascode.jira.apiTokenLogin',
     ExpandCreateWorkItemWebview = 'atlascode.jira.expandCreateWorkItem',
+    CopyImageElement = 'atlascode.jira.copyImageElement',
 
     // Debug mode-only commands
     DebugQuickCommand = 'atlascode.debug.quickCommand',
@@ -100,6 +101,7 @@ export const enum Commands {
 
     // Extension management commands
     AddRecommendedExtension = 'atlascode.addRecommendedExtension',
+    OpenRovoDevLogFile = 'OpenRovoDevLogFile',
 }
 
 // Jira projects field pagination