+++ src/container.ts
@@ -260,10 +260,14 @@ export class Container {
         context.subscriptions.push(new CustomJQLViewProvider());
         context.subscriptions.push((this._assignedWorkItemsView = new AssignedWorkItemsViewProvider()));
 
-        context.subscriptions.push(
-            (this._createWorkItemWebviewProvider = new CreateWorkItemWebviewProvider(context, context.extensionPath)),
-        );
-
+        if (this.featureFlagClient.checkGate(Features.CreateWorkItemWebviewV2)) {
+            context.subscriptions.push(
+                (this._createWorkItemWebviewProvider = new CreateWorkItemWebviewProvider(
+                    context,
+                    context.extensionPath,
+                )),
+            );
+        }
         this._onboardingProvider = new OnboardingProvider();
 
         this.refreshRovoDev(context);