+++ src/rovo-dev/rovoDevWebviewProvider.ts
@@ -1,11 +1,14 @@
 import { MinimalIssue } from '@atlassianlabs/jira-pi-common-models';
 import * as fs from 'fs';
 import path from 'path';
+import { Commands } from 'src/constants';
+import { UserInfo } from 'src/rovo-dev/api/extensionApiTypes';
 import { getFsPromise } from 'src/rovo-dev/util/fsPromises';
 import { safeWaitFor } from 'src/rovo-dev/util/waitFor';
 import { v4 } from 'uuid';
 import {
     CancellationToken,
+    commands,
     ConfigurationChangeEvent,
     Disposable,
     env,
@@ -26,6 +29,7 @@ import { GitErrorCodes } from '../typings/git';
 import { RovodevCommandContext } from './api/componentApi';
 import { DetailedSiteInfo, ExtensionApi } from './api/extensionApi';
 import { RovoDevApiClient, RovoDevHealthcheckResponse } from './client';
+import { buildErrorDetails } from './errorDetailsBuilder';
 import { RovoDevChatContextProvider } from './rovoDevChatContextProvider';
 import { RovoDevChatProvider } from './rovoDevChatProvider';
 import { RovoDevContentTracker } from './rovoDevContentTracker';
@@ -36,6 +40,7 @@ import { RovoDevProcessManager, RovoDevProcessState } from './rovoDevProcessMana
 import { RovoDevPullRequestHandler } from './rovoDevPullRequestHandler';
 import { RovoDevTelemetryProvider } from './rovoDevTelemetryProvider';
 import { RovoDevContextItem } from './rovoDevTypes';
+import { readLastNLogLines } from './rovoDevUtils';
 import {
     RovoDevDisabledReason,
     RovoDevEntitlementCheckFailedDetail,
@@ -85,6 +90,8 @@ export class RovoDevWebviewProvider extends Disposable implements WebviewViewPro
     private _debugPanelEnabled = false;
     private _debugPanelContext: Record<string, string> = {};
     private _debugPanelMcpContext: Record<string, string> = {};
+
+    private _userInfo: UserInfo | undefined;
 
     // we keep the data in this collection so we can attach some metadata to the next
     // prompt informing Rovo Dev that those files has been reverted
@@ -425,6 +432,7 @@ export class RovoDevWebviewProvider extends Disposable implements WebviewViewPro
                                 lastTenMessages: e.lastTenMessages,
                                 rovoDevSessionId: process.env.SANDBOX_SESSION_ID,
                             },
+                            this._userInfo,
                             !!this.isBoysenberry,
                         );
                         break;
@@ -470,6 +478,10 @@ export class RovoDevWebviewProvider extends Disposable implements WebviewViewPro
 
                     case RovoDevViewResponseType.OpenExternalLink:
                         await env.openExternal(Uri.parse(e.href));
+                        break;
+
+                    case RovoDevViewResponseType.OpenRovoDevLogFile:
+                        await commands.executeCommand(Commands.OpenRovoDevLogFile);
                         break;
 
                     default:
@@ -538,6 +550,8 @@ export class RovoDevWebviewProvider extends Disposable implements WebviewViewPro
                 isRetriable,
                 isProcessTerminated,
                 uid: v4(),
+                stackTrace: buildErrorDetails(error),
+                rovoDevLogs: readLastNLogLines(),
             },
         });
     }
@@ -972,6 +986,7 @@ export class RovoDevWebviewProvider extends Disposable implements WebviewViewPro
 
     private async handleProcessStateChanged(newState: RovoDevProcessState) {
         if (newState.state === 'Downloading' || newState.state === 'Starting' || newState.state === 'Started') {
+            this._userInfo = newState.jiraSiteUserInfo;
             this._jiraItemsProvider.setJiraSite(newState.jiraSiteHostname);
         }