+++ src/webviews/components/issue/create-issue-screen/CreateIssuePage.tsx
@@ -1,5 +1,4 @@
 import Button from '@atlaskit/button';
-import LoadingButton from '@atlaskit/button/loading-button';
 import Form, { ErrorMessage, Field, FormFooter, FormHeader, RequiredAsterisk } from '@atlaskit/form';
 import Page from '@atlaskit/page';
 import Select, { components } from '@atlaskit/select';
@@ -27,6 +26,7 @@ import {
     CommonEditorViewState,
     emptyCommonEditorState,
 } from '../AbstractIssueEditorPage';
+import { CreateIssueButton } from './actions/CreateIssueButton';
 import { Panel } from './Panel';
 
 type Emit = CommonEditorPageEmit;
@@ -34,6 +34,7 @@ type Accept = CommonEditorPageAccept | CreateIssueData;
 interface ViewState extends CommonEditorViewState, CreateIssueData {
     createdIssue: IssueKeyAndSite<DetailedSiteInfo>;
     formKey: string;
+    onCreateAction: 'createAndView' | 'createAndStartWork' | 'createAndGenerateCode';
 }
 
 const emptyState: ViewState = {
@@ -41,6 +42,7 @@ const emptyState: ViewState = {
     ...emptyCreateIssueData,
     createdIssue: { key: '', siteDetails: emptySiteInfo },
     formKey: v4(),
+    onCreateAction: 'createAndView',
 };
 
 const fallbackTimerDuration = 5000; // 5 seconds
@@ -108,6 +110,7 @@ export default class CreateIssuePage extends AbstractIssueEditorPage<Emit, Accep
     private attachingInProgress = false;
     private initialFieldValues: FieldValues = {};
     private suggestionFallbackTimer: NodeJS.Timeout | null = null;
+    private formRef = React.createRef<HTMLFormElement>();
 
     getProjectKey(): string {
         return this.state.fieldValues['project'].key;
@@ -236,6 +239,15 @@ export default class CreateIssuePage extends AbstractIssueEditorPage<Emit, Accep
                     });
                     break;
                 }
+                case 'createIssueWithAction': {
+                    handled = true;
+
+                    this.setState({ onCreateAction: e.action }, () => {
+                        this.formRef.current?.requestSubmit();
+                        this.setState({ onCreateAction: 'createAndView' });
+                    });
+                    break;
+                }
             }
         }
 
@@ -296,6 +308,7 @@ export default class CreateIssuePage extends AbstractIssueEditorPage<Emit, Accep
             action: 'createIssue',
             site: this.state.siteDetails,
             issueData: this.state.fieldValues,
+            onCreateAction: this.state.onCreateAction,
         };
 
         this.setState({
@@ -303,6 +316,7 @@ export default class CreateIssuePage extends AbstractIssueEditorPage<Emit, Accep
             loadingField: 'submitButton',
             lastFailedAction: createAction,
         });
+
         this.postMessage(createAction);
 
         return undefined;
@@ -313,12 +327,14 @@ export default class CreateIssuePage extends AbstractIssueEditorPage<Emit, Accep
             action: 'getScreensForSite',
             site: site,
         };
+
         this.setState({
             siteDetails: site,
             loadingField: 'site',
             isSomethingLoading: true,
             lastFailedAction: action,
         });
+
         this.postMessage(action);
     };
 
@@ -527,7 +543,7 @@ export default class CreateIssuePage extends AbstractIssueEditorPage<Emit, Accep
                                 <Form name="create-issue" key={this.state.formKey} onSubmit={this.handleSubmit}>
                                     {(frmArgs: any) => {
                                         return (
-                                            <form {...frmArgs.formProps}>
+                                            <form {...frmArgs.formProps} ref={this.formRef}>
                                                 <FormHeader title={this.formHeader()}>
                                                     <p>
                                                         Required fields are marked with an asterisk <RequiredAsterisk />
@@ -565,15 +581,18 @@ export default class CreateIssuePage extends AbstractIssueEditorPage<Emit, Accep
                                                     </Panel>
                                                 )}
                                                 <FormFooter actions={{}}>
-                                                    <LoadingButton
+                                                    <CreateIssueButton
                                                         type="submit"
-                                                        spacing="compact"
+                                                        name="Create"
                                                         className="ac-button"
-                                                        isDisabled={this.state.isSomethingLoading}
-                                                        isLoading={this.state.loadingField === 'submitButton'}
+                                                        disabled={this.state.isSomethingLoading}
+                                                        isLoading={
+                                                            this.state.isSomethingLoading &&
+                                                            this.state.loadingField === 'submitButton'
+                                                        }
                                                     >
                                                         Create
-                                                    </LoadingButton>
+                                                    </CreateIssueButton>
                                                 </FormFooter>
                                             </form>
                                         );