+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 ### [Report an Issue](https://github.com/atlassian/atlascode/issues)
 
+## What's new in 4.0.10
+
+### Features
+
+- Added development field support
+- Added Branch creation and Rovo Dev actions in Create Issue Page
+
 ## What's new in 4.0.9
 
 ### Features
@@ -12,9 +19,9 @@
 
 - Added activity history support
 
-### Bug Fixes 
+### Bug Fixes
 
-- Fixed issue with project loading in Jira Cloud Create Issue Screen 
+- Fixed issue with project loading in Jira Cloud Create Issue Screen
 
 ## What's new in 4.0.7